Moved output_redirector to utility
